% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{implied_se}
\alias{implied_se}
\title{Implied sensitivity}
\usage{
implied_se(p_obs, theta, sp)
}
\arguments{
\item{p_obs}{Observed prevalence.}

\item{theta}{True prevalence.}

\item{sp}{Specificity.}
}
\value{
Implied sensitivity.
}
\description{
Solves the forward misclassification model algebraically for Se.
}
