% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{implied_sp}
\alias{implied_sp}
\title{Implied specificity}
\usage{
implied_sp(p_obs, theta, se)
}
\arguments{
\item{p_obs}{Observed prevalence.}

\item{theta}{True prevalence.}

\item{se}{Sensitivity.}
}
\value{
Implied specificity.
}
\description{
Solves the forward misclassification model algebraically for Sp.
}
