% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcma-package.R
\docType{package}
\name{mcma-package}
\alias{mcma}
\alias{mcma-package}
\title{mcma: Bayesian Misclassification-Corrected Meta-Analysis}
\description{
Implements the Bayesian misclassification-corrected meta-analysis workflow. Wraps 'brms' to provide convenient functions for constructing corrected prevalence models, specifying sensitivity/specificity priors from external evidence, running sensitivity analyses, and estimating Se/Sp from gold-standard data via discrete kernel-weighted grid averaging (KWGA) or joint modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmfawcet/mcma}
}

}
\author{
\strong{Maintainer}: Jonathan M. Fawcett \email{jfawcett@mun.ca} (\href{https://orcid.org/0000-0002-4248-5371}{ORCID})

Authors:
\itemize{
  \item Jonathan M. Fawcett \email{jfawcett@mun.ca} (\href{https://orcid.org/0000-0002-4248-5371}{ORCID})
  \item Jedidiah Whitridge (\href{https://orcid.org/0000-0003-1237-4977}{ORCID})
}

}
\keyword{internal}
