% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mcma_sim_read}
\alias{mcma_sim_read}
\title{Read simulation results}
\usage{
mcma_sim_read(dir, pattern = "^Sim\\\\d+_\\\\d+\\\\.rds$")
}
\arguments{
\item{dir}{Directory containing simulation results.}

\item{pattern}{File pattern to match.}
}
\value{
A list of result objects.
}
\description{
Reads and combines all per-replicate RDS result files from a directory.
}
