% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwga.R
\name{print.mcma_kwga}
\alias{print.mcma_kwga}
\title{Print an mcma_kwga object}
\usage{
\method{print}{mcma_kwga}(x, ...)
}
\arguments{
\item{x}{An mcma_kwga object.}

\item{...}{Ignored.}
}
\description{
Print an mcma_kwga object
}
