% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.mcma_priors}
\alias{print.mcma_priors}
\title{Print an mcma_priors object}
\usage{
\method{print}{mcma_priors}(x, ...)
}
\arguments{
\item{x}{An mcma_priors object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print an mcma_priors object
}
