% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{print.mcma_sensitivity}
\alias{print.mcma_sensitivity}
\title{Print an mcma_sensitivity object}
\usage{
\method{print}{mcma_sensitivity}(x, ...)
}
\arguments{
\item{x}{An mcma_sensitivity object.}

\item{...}{Ignored.}
}
\description{
Print an mcma_sensitivity object
}
