% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rg_forward}
\alias{rg_forward}
\title{Forward misclassification model}
\usage{
rg_forward(theta, se, sp)
}
\arguments{
\item{theta}{True prevalence (numeric vector).}

\item{se}{Sensitivity (numeric vector).}

\item{sp}{Specificity (numeric vector).}
}
\value{
Expected observed prevalence.
}
\description{
Returns the expected observed prevalence given true prevalence, Se, and Sp.
}
