% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwga.R
\name{summary.mcma_kwga}
\alias{summary.mcma_kwga}
\title{Summary for mcma_kwga}
\usage{
\method{summary}{mcma_kwga}(object, ...)
}
\arguments{
\item{object}{An mcma_kwga object.}

\item{...}{Ignored.}
}
\description{
Summary for mcma_kwga
}
